♻️ Refactor Edit/Add product pages
🏗️ Move form from "EditProduct" to "EditProductForm" component.
📄 Create helpers.js file with common data.
🏗️ Move: (sizesArray, imagesArray, countOfPictures) from "Add/Edit product" pages to helpers.
